+++ pyrexplorer/__init__.py
@@ -14,5 +14,5 @@ in analyzed data (patterns discovery, i.e. data mining).
 
 __author__ = 'Mikhail Titov'
 __contact__ = 'mikhail.titov@cern.ch'
-__version__ = '0.2.1'
-__date__ = '2015-04-27'
+__version__ = '0.3.2'
+__date__ = '2015-05-27'